+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Remove AccountsUI.preferredLabel() method, moved to pwix:accounts-tools
     - Make sure old and new passwords are different when changing a password
     - Have a 'data-ac-name' attribute on acUserLogin (todo #20)
+    - Remove save() and restore() methods (obsoleting todo #55)
 
 ### 1.3.0